+++ terraform/variables.tf
@@ -18,3 +18,9 @@ variable "namespace" {
   description = "Namespace to deploy proxies etc. in to, for canaries or deploys. To make it prettier, start with a hyphen (e.g. '-apm-123')."
   default = ""
 }
+
+variable "force_sandbox" {
+  type = bool
+  description = "Force a sandbox deploy instead of trying to detect if the deploy is happenning in a sandbox env"
+  default = false
+}